refactor(android_native): add SherpaOnnx STT and TTS helper classes for speech processing
